# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API documentation",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API server"
        }
    ],
    "paths": {
...
        "/v1/register": {
            "post": {
                "tags": [
                    "Авторизация/регистрация"
                ],
                "summary": "Регистрация нового пользователя",
                "operationId": "8d19ebf27aad77f03d818b41eb43fbf1",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "type",
                                    "inn",
                                    "phone",
                                    "form"
                                ],
                                "properties": {
                                    "type": {
                                        "type": "string",
                                        "enum": [
                                            "Ритуальное агенство",
                                            "Частный агент",
                                            "Другое"
                                        ],
                                        "example": "Ритуальное агенство"
                                    },
                                    "inn": {
                                        "type": "string",
                                        "maxLength": 12,
                                        "example": "123456789012"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "maxLength": 15,
                                        "example": "+79991234567"
                                    },
                                    "form": {
                                        "type": "string",
                                        "enum": [
                                            "ООО",
                                            "ИП",
                                            "ФЛ"
                                        ],
                                        "example": "ООО"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Успешная регистрация",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "reg_id": {
                                            "type": "string",
                                            "format": "uuid",
                                            "example": "550e8400-e29b-41d4-a716-446655440000"
                                        },
                                        "agent_name": {
                                            "type": "string",
                                            "example": "ООО Ромашка"
                                        },
                                        "status": {
                                            "type": "string",
                                            "example": "ACTIVE"
                                        },
                                        "okved": {
                                            "type": "string",
                                            "example": "62.01"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Ошибка валидации",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "object",
                                            "example": {
                                                "type": [
                                                    "Поле type обязательно для заполнения."
                                                ]
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Пользователь уже существует или ИНН не найден",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Пользователь с таким номером телефона или инн уже существует."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/v1/auth": {
            "post": {
                "tags": [
                    "Авторизация/регистрация"
                ],
                "summary": "Инициализация авторизации",
                "operationId": "dc43034ebd4bec1ae90f0cd3170d4ae8",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "phone"
                                ],
                                "properties": {
                                    "phone": {
                                        "type": "string",
                                        "example": "+79991234567"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "SMS отправлен",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "reg_id": {
                                            "type": "string",
                                            "format": "uuid",
                                            "example": "550e8400-e29b-41d4-a716-446655440000"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Ошибка валидации",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "object",
                                            "example": {
                                                "phone": [
                                                    "Поле phone обязательно для заполнения."
                                                ]
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "description": "Введите токен без префикса Bearer"
            }
        }
    },
...
    "security": [
        {
            "bearerAuth": []
        }
    ]
}